Adding detailed diffs (#5255)
* Added diff implementation from mocha

Moved logic to separate files

* Removed IE-specific logic

* Refactoring, separated diff-stringify

* Fixed lint

* Removed changes to ./package.json

* .js -> .ts; not fully working yet

* Refactoring

* Added forgotten files to git

* Fixed server tests

* Fixed error messages and some functional tests

* Fixed functional tests

* Fixed lint

* Removed 'index' from import

* Refactored stringify() to use util.inspect instead of JSON.stringify

* Integrated colors with the existing system

* Fixed functional tests; generateDiff -> generate; removed stringify export

* Using debug logger; Had to refactor some errors into ./src/errors/test-run/shared.js

* Importing only the required functions from external modules

* Added server test for different object types. Found a bug, will fix in the next commit

* Fixed the comma bug

* Fixed server tests

* Rebased onto master, fixed errors

* Fixing tests - changed package.json

* Removed node-polyfills plugin

* Moved colors to constant

* Fixed codestyle (aside from shared.js filename)

* Moved shared errors into src/shared/errors; refactored cleanUpFilter
