+++ package.json
@@ -78,6 +78,7 @@
     "dedent": "^0.4.0",
     "del": "^3.0.0",
     "device-specs": "^1.0.0",
+    "diff": "^4.0.2",
     "elegant-spinner": "^1.0.1",
     "emittery": "^0.4.1",
     "endpoint-utils": "^1.0.2",
@@ -139,6 +140,7 @@
     "@types/chai": "^3.5.2",
     "@types/debug": "^4.1.5",
     "@types/dedent": "^0.7.0",
+    "@types/diff": "^4.0.2",
     "@types/json5": "0.0.30",
     "@types/mime-db": "^1.27.0",
     "@types/mustache": "^0.8.32",
