+++ src/client/driver/command-executors/client-functions/client-function-executor.js
@@ -2,7 +2,7 @@ import { Promise } from '../../deps/hammerhead';
 import DriverStatus from '../../status';
 import { createReplicator, FunctionTransform, ClientFunctionNodeTransform } from './replicator';
 import evalFunction from './eval-function';
-import { UncaughtErrorInClientFunctionCode } from '../../../../errors/test-run';
+import { UncaughtErrorInClientFunctionCode } from '../../../../shared/errors';
 
 export default class ClientFunctionExecutor {
     constructor (command) {
