+++ test/functional/fixtures/api/es-next/generic-errors/test.js
@@ -52,19 +52,19 @@ describe('[API] Generic errors', function () {
         it('Should handle Node built-in assertion lib error', function () {
             const NODE_11_ASSERTION_MESSAGE = [
                 'AssertionError [ERR_ASSERTION]: Expected values to be strictly equal:',
-                '+ actual - expected',
-                '',
+                '+ actual',
+                '- expected',
                 '+ \'answer\'',
                 '- \'42\''
-            ].join(' ');
+            ];
 
             const NODE_10_ASSERTION_MESSAGE = [
                 'AssertionError [ERR_ASSERTION]: Input A expected to strictly equal input B:',
-                '+ expected - actual',
-                '',
+                '+ expected',
+                '- actual',
                 '- \'answer\'',
                 '+ \'42\''
-            ].join(' ');
+            ];
 
             const OLD_NODE_ASSERTION_MESSAGE_RE = /AssertionError( \[ERR_ASSERTION])?: 'answer' === '42'/;
 
@@ -73,10 +73,16 @@ describe('[API] Generic errors', function () {
                 .catch(function (errs) {
                     expect(errs[0]).to.contains('> 13 |    assert.strictEqual(\'answer\', \'42\');');
 
-                    if (nodeVersion.major >= 11)
-                        expect(errs[0]).to.contain(NODE_11_ASSERTION_MESSAGE);
-                    else if (nodeVersion.major >= 10)
-                        expect(errs[0]).to.contain(NODE_10_ASSERTION_MESSAGE);
+                    if (nodeVersion.major >= 11) {
+                        NODE_11_ASSERTION_MESSAGE.forEach((item) => {
+                            expect(errs[0]).to.contain(item);
+                        });
+                    }
+                    else if (nodeVersion.major >= 10) {
+                        NODE_10_ASSERTION_MESSAGE.forEach((item) => {
+                            expect(errs[0]).to.contain(item);
+                        });
+                    }
                     else
                         expect(errs[0]).to.match(OLD_NODE_ASSERTION_MESSAGE_RE);
                 });
