+++ test/functional/fixtures/test-run/testcafe-fixtures/get-real-queue-length.js
@@ -1,6 +1,5 @@
-import { PressKeyCommand } from '../../../../../lib/test-run/commands/actions';
+import { PressKeyCommand, ExecuteAsyncExpressionCommand, ExecuteExpressionCommand } from '../../../../../lib/test-run/commands/actions';
 import { WaitCommand } from '../../../../../lib/test-run/commands/observation';
-import { ExecuteAsyncExpressionCommand, ExecuteExpressionCommand } from '../../../../../src/test-run/commands/actions';
 
 
 fixture `Test run commands queue`
