+++ test/server/data/expected-test-run-errors/external-assertion-library-errors/array
@@ -1,0 +1,39 @@
+AssertionError: expected [ 1, 2, 3, [ 4, 5 ], 6 ] to deeply equal [ 1, 4, 2,
+3, [ 5, 6, [ 7 ] ] ]
+
++ expected - actual
+
+ [
+   1
++  4
+   2
+   3
+   [
+-    4
+     5
++    6
++    [
++      7
++    ]
+   ]
+-  6
+ ]
+
+
+Browser: Chrome 15.0.874.120 / macOS 10.15
+Screenshot: /unix/path/with/<tag>
+
+   18 |function func1 () {
+   19 |    record = createCallsiteRecord({ byFunctionName: 'func1' });
+   20 |}
+   21 |
+   22 |(function func2 () {
+ > 23 |    func1();
+   24 |})();
+   25 |
+   26 |stackTrace.filter.deattach(stackFilter);
+   27 |
+   28 |module.exports = record;
+
+   at func2 (testfile.js:23:5)
+   at Object.<anonymous> (testfile.js:24:3)
