+++ test/server/data/expected-test-run-errors/external-assertion-library-errors/number
@@ -1,4 +1,10 @@
-AssertionError: expected true to deeply equal false
+AssertionError: expected 1 to deeply equal 2
+
++ expected - actual
+
+-1
++2
+
 
 Browser: Chrome 15.0.874.120 / macOS 10.15
 Screenshot: /unix/path/with/<tag>
